pyproject.toml
[tool.poetry]
name = "nonebot-plugin-7s-roll"
version = "0.2.0a1"
description = "A roll dice plugin for nonebot"
authors = ["7sDream <i@7sdre.am>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/7sDream/nonebot_plugin_7s_roll"
repository = "https://github.com/7sDream/nonebot_plugin_7s_roll"
keywords = ["nonebot", "dice", "roll", "trpg"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Plugins",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
include = [
    "LICENSE",
]

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"
black = "^23.3.0"
rope = "^1.8.0"
pytest = "^7.3.2"
nonebot-adapter-console = "^0.3.2"
nonebot2 = {version = "^2.0.0", extras = ["fastapi"]}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
